crnn/requirements.txt
lmdb==1.3.0
numpy==1.19.3
pandas==1.3.1
Pillow==9.0.1
six==1.16.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.27.0
